project/AuthenticationServer/src/bg/sofia/uni/fmi/mjt/authserver/user/PasswordHasher.java
package bg.sofia.uni.fmi.mjt.authserver.user;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;

public class PasswordHasher {
    private static final int PAD = 0xff;
    public static String hash(String input) {
        if (input == null) {
            throw new IllegalArgumentException("input cannot be null");
        }

        try {
            MessageDigest digest = MessageDigest.getInstance("SHA-256");
            byte[] hashedBytes = digest.digest(input.getBytes());

            StringBuilder hexString = new StringBuilder();
            for (byte b : hashedBytes) {
                String hex = Integer.toHexString(PAD & b);
                if (hex.length() == 1) {
                    hexString.append('0');
                }
                hexString.append(hex);
            }
            return hexString.toString();
        } catch (NoSuchAlgorithmException e) {
            throw new RuntimeException(e);
        }
    }
}
